Moved discord signup to its own page
